+++ src/extension.ts
@@ -1,25 +1,130 @@
-// The module 'vscode' contains the VS Code extensibility API
-// Import the module and reference it with the alias vscode in your code below
-import * as vscode from 'vscode';
+import { Credentials } from "./GitHub/authentication";
+import * as config from "./config";
+import * as trace from "./tracing";
+import { commands, ExtensionContext, workspace, window } from "vscode";
+import { GistProvider } from "./Tree/nodes";
+import { GistFileSystemProvider, GIST_SCHEME } from "./FileSystem/fileSystem";
+import { TGitHubUser } from "./GitHub/types";
+import { clearGlobalStorage, getFollowedUsersFromGlobalStorage,  removeFromGlobalStorage } from "./FileSystem/storage";
+import { GLOBAL_STORAGE_KEY } from "./GitHub/constants";
+import { getGitHubAuthenticatedUser } from "./GitHub/api";
 
-// this method is called when your extension is activated
-// your extension is activated the very first time the command is executed
-export function activate(context: vscode.ExtensionContext) {
-	
-	// Use the console to output diagnostic information (console.log) and errors (console.error)
-	// This line of code will only be executed once when your extension is activated
-	console.log('Congratulations, your extension "vscode-Gists" is now active!');
+export let output: trace.Output;
+export const credentials = new Credentials();
+export let gitHubAuthenticatedUser: TGitHubUser;
+export let extensionContext: ExtensionContext;
+export const gistProvider = new GistProvider();
+export const gistFileSystemProvider = new GistFileSystemProvider();
 
-	// The command has been defined in the package.json file
-	// Now provide the implementation of the command with registerCommand
-	// The commandId parameter must match the command field in package.json
-	let disposable = vscode.commands.registerCommand('vscode-Gists.helloWorld', () => {
-		// The code you place here will be executed every time your command is executed
-		// Display a message box to the user
-		vscode.window.showInformationMessage('Hello World from Gists for Visual Studio Code!');
-	});
+export async function activate(context: ExtensionContext) {
+    extensionContext = context;
+    if (config.get("EnableTracing")) {
+        output = new trace.Output();
+    }
 
-	context.subscriptions.push(disposable);
+    gitHubAuthenticatedUser = await getGitHubAuthenticatedUser();
+
+    output?.appendLine("Virtual Gists: extension is now active!", output.messageType.info);
+
+    await credentials.initialize(context);
+    if (!credentials.isAuthenticated) {
+        credentials.initialize(context);
+    }
+    const disposable = commands.registerCommand("extension.getGitHubUser", async () => {
+        const octokit = await credentials.getOctokit();
+        const userInfo = await octokit.users.getAuthenticated();
+
+        output?.appendLine(`Logged to GitHub as ${userInfo.data.login}`, output.messageType.info);
+    });
+
+    context.subscriptions.push(
+        commands.registerCommand("VirtualGists.refreshTree", async () => {
+            gistProvider.refresh();
+        })
+    );
+
+    context.subscriptions.push(
+        commands.registerCommand("VirtualGists.getGlobalStorage", async () => {
+            const reposFromGlobalStorage = await getFollowedUsersFromGlobalStorage(context);
+            if (reposFromGlobalStorage.length > 0) {
+                output?.appendLine(`Global storage: ${reposFromGlobalStorage}`, output.messageType.info);
+            } else {
+                output?.appendLine(`Global storage is empty`, output.messageType.info);
+            }
+        })
+    );
+
+    context.subscriptions.push(
+        commands.registerCommand("VirtualGists.purgeGlobalStorage", async () => {
+            // purgeGlobalStorage(extensionContext);
+            throw new Error("Not implemented");
+        })
+    );
+
+    context.subscriptions.push(
+        commands.registerCommand("VirtualGists.removeFromGlobalStorage", async () => {
+            const gistsFromGlobalStorage = await getFollowedUsersFromGlobalStorage(context);
+            const gistToRemove = await window.showQuickPick(gistsFromGlobalStorage, {
+                placeHolder: "Select gist to remove from global storage",
+                ignoreFocusOut: true,
+                canPickMany: false,
+            });
+            if (gistToRemove) {
+                removeFromGlobalStorage(context, gistToRemove);
+            }
+        })
+    );
+
+    context.subscriptions.push(
+        commands.registerCommand("VirtualGists.clearGlobalStorage", async () => {
+            clearGlobalStorage(context);
+        })
+    );
+
+    context.subscriptions.push(
+        workspace.registerFileSystemProvider(GIST_SCHEME, gistFileSystemProvider, {
+            isCaseSensitive: true,
+        })
+    );
+
+    // context.subscriptions.push(
+    //     commands.registerCommand("VirtualRepos.openRepository", async () => {
+    //         const pick = (await pickRepository()) as string;
+    //         if (pick) {
+    //             output?.appendLine(`Picked repository: ${pick}`, output.messageType.info);
+    //             await addToGlobalStorage(context, pick);
+    //         } else {
+    //             output?.appendLine("Open repository cancelled by uer", output.messageType.info);
+    //         }
+    //     })
+    // );
+
+    // register global storage
+    const keysForSync = [GLOBAL_STORAGE_KEY];
+    context.globalState.setKeysForSync(keysForSync);
+
+    context.subscriptions.push(
+        workspace.onDidChangeConfiguration((e) => {
+            if (e.affectsConfiguration("VirtualGists.EnableTracing")) {
+                if (config.get("EnableTracing")) {
+                    output = new trace.Output();
+                } else {
+                    output?.dispose();
+                }
+            }
+        })
+    );
+
+    window.createTreeView("virtualGistsView", {
+        treeDataProvider: gistProvider,
+        showCollapseAll: true,
+        canSelectMany: true,
+    });
+    // tv.reveal(store.repos);
+
+    // window.registerTreeDataProvider("Repositories", repoProvider);
+
+    context.subscriptions.push(disposable);
 }
 
 // this method is called when your extension is deactivated